Add @vercel/ncc package and create 'package' script
The @vercel/ncc package was added to the dependencies in package.json and package-lock.json. This package will be used to compile Node.js applications into a single file. Along with the addition of this package, a 'package' script has been added to the scripts section of package.json to facilitate the bundling of the application.

Took 10 minutes