Add 'skipErrorChecking' input variable and update README
The code introduces an optional input 'skipErrorChecking' to control TypeDoc error checking. Its effect has been explained in the README. This upgrade provides users with the choice to forgo type checking, possibly improving generation time, but should be used cautiously as it may lead to crashes if the code contains type errors.